src/SampleProject/ClientApp/src/app/entities/task-item/route/task-item-routing-resolve.service.spec.ts
import { TestBed } from "@angular/core/testing";
import { HttpResponse } from "@angular/common/http";
import { HttpClientTestingModule } from "@angular/common/http/testing";
import {
  ActivatedRouteSnapshot,
  ActivatedRoute,
  Router,
  convertToParamMap,
} from "@angular/router";
import { RouterTestingModule } from "@angular/router/testing";
import { of } from "rxjs";

import { ITaskItem, TaskItem } from "../task-item.model";
import { TaskItemService } from "../service/task-item.service";

import { TaskItemRoutingResolveService } from "./task-item-routing-resolve.service";

describe("TaskItem routing resolve service", () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TaskItemRoutingResolveService;
  let service: TaskItemService;
  let resultTaskItem: ITaskItem | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest
      .spyOn(mockRouter, "navigate")
      .mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(TaskItemRoutingResolveService);
    service = TestBed.inject(TaskItemService);
    resultTaskItem = undefined;
  });

  describe("resolve", () => {
    it("should return ITaskItem returned by find", () => {
      // GIVEN
      service.find = jest.fn((id) => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService
        .resolve(mockActivatedRouteSnapshot)
        .subscribe((result) => {
          resultTaskItem = result;
        });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTaskItem).toEqual({ id: 123 });
    });

    it("should return new ITaskItem if id is not provided", () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService
        .resolve(mockActivatedRouteSnapshot)
        .subscribe((result) => {
          resultTaskItem = result;
        });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTaskItem).toEqual(new TaskItem());
    });

    it("should route to 404 page if data not found in server", () => {
      // GIVEN
      jest
        .spyOn(service, "find")
        .mockReturnValue(
          of(new HttpResponse({ body: null as unknown as TaskItem }))
        );
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService
        .resolve(mockActivatedRouteSnapshot)
        .subscribe((result) => {
          resultTaskItem = result;
        });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTaskItem).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(["404"]);
    });
  });
});
